+++ packages/ai/core/generate-image/generate-image-result.ts
@@ -1,3 +1,4 @@
+import { GeneratedFile } from '../generate-text';
 import { ImageGenerationWarning } from '../types/image-model';
 import { ImageModelResponseMetadata } from '../types/image-model-response-metadata';
 
@@ -9,12 +10,12 @@ export interface GenerateImageResult {
   /**
 The first image that was generated.
    */
-  readonly image: GeneratedImage;
+  readonly image: GeneratedFile;
 
   /**
 The images that were generated.
      */
-  readonly images: Array<GeneratedImage>;
+  readonly images: Array<GeneratedFile>;
 
   /**
 Warnings for the call, e.g. unsupported settings.
@@ -25,16 +26,4 @@ Warnings for the call, e.g. unsupported settings.
 Response metadata from the provider. There may be multiple responses if we made multiple calls to the model.
    */
   readonly responses: Array<ImageModelResponseMetadata>;
-}
-
-export interface GeneratedImage {
-  /**
-Image as a base64 encoded string.
-   */
-  readonly base64: string;
-
-  /**
-Image as a Uint8Array.
-   */
-  readonly uint8Array: Uint8Array;
 }