+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,20 @@
 
 ### Patch Changes
 
+- Updated dependencies [0bd5bc6]
+  - ai@4.1.63
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [c9ed3c4]
+  - ai@4.1.62
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [2e1101a]
   - ai@4.1.61