+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/azure
 
+## 1.2.6
+
+### Patch Changes
+
+- Updated dependencies [0bd5bc6]
+  - @ai-sdk/provider@1.0.12
+  - @ai-sdk/openai@1.2.6
+  - @ai-sdk/provider-utils@2.1.14
+
 ## 1.2.5
 
 ### Patch Changes